+++ cmd/notation/list.go
@@ -3,9 +3,10 @@ package main
 import (
 	"context"
 	"fmt"
+
 	notationregistry "github.com/notaryproject/notation-go/registry"
+	"github.com/notaryproject/notation/cmd/notation/internal/cmdutil"
 	"github.com/notaryproject/notation/cmd/notation/internal/experimental"
-	"github.com/notaryproject/notation/internal/arg"
 	"github.com/notaryproject/notation/internal/cmd"
 	"github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
@@ -32,7 +33,7 @@ func listCommand(opts *listOpts) *cobra.Command {
 		Short:   "List signatures of the signed artifact",
 		Long:    "List all the signatures associated with signed artifact",
 		Args: func(cmd *cobra.Command, args []string) error {
-			if err := arg.ValidateCount(args, 1, "missing reference"); err != nil {
+			if err := cmdutil.ValidateArgsCount(args, 1, "missing reference"); err != nil {
 				return err
 			}
 			opts.reference = args[0]
